✨ feat: App 컨트롤러 추가
