+++ index.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>점심 뭐 먹지</title>
   </head>
-  <body>
+  <body id="app">
     <!-- GNB -->
     <header class="gnb"></header>
